refactor: move OnspringRequest under fluent namespace
