+++ Onspring.API.SDK.Tests/Tests/OnspringClientFluentTests.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Onspring.API.SDK.Models;
+using Onspring.API.SDK.Models.Fluent;
 
 namespace Onspring.API.SDK.Tests.Tests
 {
