+++ Onspring.API.SDK.Tests/Tests/OnspringRequestTests.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Onspring.API.SDK.Models;
 
 namespace Onspring.API.SDK.Tests.Tests
 {
