+++ Onspring.API.SDK/OnspringClient.cs
@@ -3,6 +3,7 @@ using Onspring.API.SDK.Extensions;
 using Onspring.API.SDK.Helpers;
 using Onspring.API.SDK.Http;
 using Onspring.API.SDK.Models;
+using Onspring.API.SDK.Models.Fluent;
 using Onspring.API.SDK.Validation;
 using System;
 using System.Collections.Generic;
